fix SP1_groth16 files
