+++ Makefile
@@ -243,8 +243,8 @@ build-sp1-macos:
 
 build-sp1-groth16-macos:
 	@cd operator/sp1_groth16/lib && cargo build --release
-	@cp operator/sp1_groth16/lib/target/release/libsp1_verifier_ffi.dylib operator/sp1_groth16/lib/libsp1_verifier.dylib
-	@cp operator/sp1_groth16/lib/target/release/libsp1_verifier_ffi.a operator/sp1_groth16/lib/libsp1_verifier.a
+	@cp operator/sp1_groth16/lib/target/release/libsp1_verifier_ffi.dylib operator/sp1_groth16/lib/libsp1_groth16_verifier.dylib
+	@cp operator/sp1_groth16/lib/target/release/libsp1_verifier_ffi.a operator/sp1_groth16/lib/libsp1_groth16_verifier.a
