--- operator/sp1_groth16/lib/sp1.h
@@ -1,4 +1,0 @@
-#include <stdbool.h>
-
-bool verify_sp1_proof_ffi(unsigned char *proof_buffer, unsigned int proof_len,
-                          unsigned char *elf_buffer, unsigned int elf_len);
