+++ operator/sp1_groth16/lib/src/lib.rs
@@ -3,7 +3,7 @@ pub const MAX_PROOF_SIZE: usize = 2 * 1024 * 1024;
 pub const MAX_ELF_BUFFER_SIZE: usize = 1024 * 1024;
 
 #[no_mangle]
-pub extern "C" fn verify_sp1_proof_ffi(
+pub extern "C" fn verify_sp1_groth16_proof_ffi(
     proof_bytes: &[u8; MAX_PROOF_SIZE],
     proof_len: usize,
     elf_bytes: &[u8; MAX_ELF_BUFFER_SIZE],
@@ -30,7 +30,7 @@ mod tests {
         include_bytes!("../../../../task_sender/test_examples/sp1_groth16/elf/riscv32im-succinct-zkvm-elf");
 
     #[test]
-    fn verify_sp1_proof_with_elf_works() {
+    fn verify_sp1_groth16_proof_with_elf_works() {
         let mut proof_buffer = [0u8; MAX_PROOF_SIZE];
         let proof_size = PROOF.len();
         proof_buffer[..proof_size].clone_from_slice(PROOF);
@@ -39,12 +39,12 @@ mod tests {
         let elf_size = ELF.len();
         elf_buffer[..elf_size].clone_from_slice(ELF);
 
-        let result = verify_sp1_proof_ffi(&proof_buffer, proof_size, &elf_buffer, elf_size);
+        let result = verify_sp1_groth16_proof_ffi(&proof_buffer, proof_size, &elf_buffer, elf_size);
         assert!(result)
     }
 
     #[test]
-    fn verify_sp1_aborts_with_bad_proof() {
+    fn verify_sp1_groth16_aborts_with_bad_proof() {
         let mut proof_buffer = [42u8; super::MAX_PROOF_SIZE];
         let proof_size = PROOF.len();
         proof_buffer[..proof_size].clone_from_slice(PROOF);
@@ -53,7 +53,7 @@ mod tests {
         let elf_size = ELF.len();
         elf_buffer[..elf_size].clone_from_slice(ELF);
 
-        let result = verify_sp1_proof_ffi(&proof_buffer, proof_size - 1, &elf_buffer, elf_size);
+        let result = verify_sp1_groth16_proof_ffi(&proof_buffer, proof_size - 1, &elf_buffer, elf_size);
         assert!(!result)
     }
 }
